+++ src/main/java/com/example/document_flow/repository/DAO/DocumentDAO.java
@@ -17,5 +17,5 @@ public interface DocumentDAO extends DAO<Document> {
      * @param id автора документов
      * @return перечень документов, созданных автором с указанным id
      */
-    List<Document> getDocumentByIdAuthor(int id);
+    List<Document> getDocumentByIdAuthor(long id);
 }
